+++ DeanzaClassRegister/DeanzaClassRegister/SectionInfo.swift
@@ -1,0 +1,14 @@
+//
+//  SectionInfo.swift
+//  DeanzaClassRegister
+//
+//  Created by Alvin Lin on 2018/8/20.
+//  Copyright © 2018 Alvin Lin. All rights reserved.
+//
+
+import Foundation
+
+struct SectionInfo {
+    var departmentList: [String]
+    var isExpanded: [Bool]
+}
